Move generator/parser unit tests to new locations
In converterTest, translatorTest, and inchiTest
